testLogBox package: created sub-packages and moved tests ; POM file:  add profile; Add testng.xml file
