+++ LogBox20220513/src/test/java/testsLogBox/ActivityFeedTests/ActivityTests.java
@@ -1,12 +1,10 @@
-package testsLogBox;
+package testsLogBox.ActivityFeedTests;
 
 import static org.testng.Assert.assertTrue;
 
 import java.io.IOException;
 
 import org.hamcrest.core.StringContains;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.Reporter;
 import org.testng.annotations.Test;
@@ -15,21 +13,20 @@ import frameWork.BasePageFrameWork;
 import pageObjectsLogBox.PageObjectsBrochurePage;
 import pageObjectsLogBox.PageObjectsHomePage;
 
-public class HomePageTests extends BasePageFrameWork{
+public class ActivityTests extends BasePageFrameWork{
 	
 	PageObjectsHomePage pageObjectsHomePage = new PageObjectsHomePage();
 	PageObjectsBrochurePage pageObjectsBrochurePage = new PageObjectsBrochurePage();
 	
-	@Test
 	public void shouldSearchForPatient() throws InterruptedException, IOException {
 		pageObjectsBrochurePage.selectPracticeAndClickLoginButton();
 		pageObjectsBrochurePage.insertUsernameAndPasswordFromExcel(1,1,2);
     	pageObjectsBrochurePage.clickLoginButtonToSubmitUsernameAndPassword();
 		pageObjectsHomePage.searchPracticePatientsOnHomePage("John");
-		WebElement searchedPatient = driver.findElement(By.cssSelector("[data-cy] tr td:nth-of-type(2)"));
-		//searchedPatient.toString();
-		//Assert.assertTrue(searchedPatient.contains("John"));
-		
+		pageObjectsHomePage.clickOnSearchedPatient();
+		String feedURL = driver.getCurrentUrl();
+		Assert.assertTrue(feedURL.contains("feed"));
+		Reporter.log("User ");
 	}
-	
+
 }
