+++ LogBox20220513/src/test/java/testsLogBox/AdmissionTests/AdmissionTests.java
@@ -1,4 +1,4 @@
-package testsLogBox;
+package testsLogBox.AdmissionTests;
 
 import frameWork.BasePageFrameWork;
