+++ LogBox20220513/src/test/java/testsLogBox/DischargesTests/DischargesTests.java
@@ -1,4 +1,4 @@
-package testsLogBox;
+package testsLogBox.DischargesTests;
 
 import frameWork.BasePageFrameWork;
