+++ LogBox20220513/src/test/java/testsLogBox/HomePageTests/HomePageTests.java
@@ -1,10 +1,12 @@
-package testsLogBox;
+package testsLogBox.HomePageTests;
 
 import static org.testng.Assert.assertTrue;
 
 import java.io.IOException;
 
 import org.hamcrest.core.StringContains;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.Reporter;
 import org.testng.annotations.Test;
@@ -13,20 +15,21 @@ import frameWork.BasePageFrameWork;
 import pageObjectsLogBox.PageObjectsBrochurePage;
 import pageObjectsLogBox.PageObjectsHomePage;
 
-public class ActivityTests extends BasePageFrameWork{
+public class HomePageTests extends BasePageFrameWork{
 	
 	PageObjectsHomePage pageObjectsHomePage = new PageObjectsHomePage();
 	PageObjectsBrochurePage pageObjectsBrochurePage = new PageObjectsBrochurePage();
 	
+	@Test
 	public void shouldSearchForPatient() throws InterruptedException, IOException {
 		pageObjectsBrochurePage.selectPracticeAndClickLoginButton();
 		pageObjectsBrochurePage.insertUsernameAndPasswordFromExcel(1,1,2);
     	pageObjectsBrochurePage.clickLoginButtonToSubmitUsernameAndPassword();
 		pageObjectsHomePage.searchPracticePatientsOnHomePage("John");
-		pageObjectsHomePage.clickOnSearchedPatient();
-		String feedURL = driver.getCurrentUrl();
-		Assert.assertTrue(feedURL.contains("feed"));
-		Reporter.log("User ");
+		WebElement searchedPatient = driver.findElement(By.cssSelector("[data-cy] tr td:nth-of-type(2)"));
+		//searchedPatient.toString();
+		//Assert.assertTrue(searchedPatient.contains("John"));
+		
 	}
-
+	
 }
