+++ LogBox20220513/src/test/java/testsLogBox/LoginTests/LoginPracticeTests.java
@@ -1,4 +1,4 @@
-package testsLogBox;
+package testsLogBox.LoginTests;
 import java.io.IOException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
