+++ LogBox20220513/test-output/Charlotte/Pre-admission, Admission and Discharge.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-<title>TestNG:  Default test</title>
+<title>TestNG:  Pre-admission, Admission and Discharge</title>
 <link href="../testng.css" rel="stylesheet" type="text/css" />
 <link href="../my-testng.css" rel="stylesheet" type="text/css" />
 
@@ -53,13 +53,13 @@ function toggleAllBoxes() {
 
 </head>
 <body>
-<h2 align='center'>Default test</h2><table border='1' align="center">
+<h2 align='center'>Pre-admission, Admission and Discharge</h2><table border='1' align="center">
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Wed Jun 01 08:33:36 SAST 2022</td>
+<td>Started on:</td><td>Wed Jun 01 08:44:00 SAST 2022</td>
 </tr>
-<tr><td>Total time:</td><td>9 seconds (9900 ms)</td>
+<tr><td>Total time:</td><td>3 seconds (3217 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,10 +75,17 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='testsLogBox.HomePageTests.HomePageTests.shouldSearchForPatient()'><b>shouldSearchForPatient</b><br>Test class: testsLogBox.HomePageTests.HomePageTests</td>
+<td title='testsLogBox.LoginTests.LoginPracticeTests.shouldLoginWithValidUsernameValidPassword()'><b>shouldLoginWithValidUsernameValidPassword</b><br>Test class: testsLogBox.LoginTests.LoginPracticeTests<br/>
+<a href="#Output-640161448" onClick='toggleBox("Output-640161448", this, "Show output", "Hide output");'>Show output</a>
+
+<a href="#Output-640161448" onClick="toggleAllBoxes();">Show all outputs</a>
+<div class='log' id="Output-640161448">
+Logged in with a valid username and valid password into Expected URL = https://qa.logbox.co.za/premium/#/<br/>
+</div>
+</td>
 <td></td>
-<td>9</td>
-<td>testsLogBox.HomePageTests.HomePageTests@40021799</td></tr>
+<td>3</td>
+<td>testsLogBox.LoginTests.LoginPracticeTests@81ff872</td></tr>
 </table><p>
 </body>
 </html>
