+++ LogBox20220513/test-output/Charlotte/Pre-admission, Admission and Discharge.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="YolandeH-Dell2" failures="0" tests="1" name="Pre-admission, Admission and Discharge" time="3.217" errors="0" timestamp="01 Jun 2022 06:44:03 GMT">
+  <testcase classname="testsLogBox.LoginTests.LoginPracticeTests" name="shouldLoginWithValidUsernameValidPassword" time="3.195"/>
+</testsuite> <!-- Pre-admission, Admission and Discharge -->
