+++ LogBox20220513/test-output/Default suite/Default test.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="YolandeH-Dell2" failures="0" tests="1" name="Default test" time="3.73" errors="0" timestamp="28 May 2022 10:47:21 GMT">
-  <testcase classname="testsLogBox.HomePageTests" name="shouldSearchForPatient" time="3.71"/>
+<testsuite ignored="0" hostname="YolandeH-Dell2" failures="0" tests="1" name="Default test" time="9.9" errors="0" timestamp="01 Jun 2022 06:33:46 GMT">
+  <testcase classname="testsLogBox.HomePageTests.HomePageTests" name="shouldSearchForPatient" time="9.882"/>
 </testsuite> <!-- Default test -->
