+++ LogBox20220513/test-output/emailable-report.html
@@ -8,11 +8,11 @@
 <body>
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
-<tr><th colspan="7">Default suite</th></tr>
-<tr><td><a href="#t0">Default test</a></td><td class="num">1</td><td class="num">0</td><td class="num">0</td><td class="num">3 730</td><td></td><td></td></tr>
+<tr><th colspan="7">Charlotte</th></tr>
+<tr><td><a href="#t0">Pre-admission, Admission and Discharge</a></td><td class="num">1</td><td class="num">0</td><td class="num">0</td><td class="num">3 217</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">testsLogBox.HomePageTests</td><td><a href="#m0">shouldSearchForPatient</a></td><td rowspan="1">1653734838111</td><td rowspan="1">3710</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Charlotte</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Pre-admission, Admission and Discharge &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">testsLogBox.LoginTests.LoginPracticeTests</td><td><a href="#m0">shouldLoginWithValidUsernameValidPassword</a></td><td rowspan="1">1654065840220</td><td rowspan="1">3195</td></tr></tbody>
 </table>
-<h2>Default test</h2><h3 id="m0">testsLogBox.HomePageTests#shouldSearchForPatient</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Pre-admission, Admission and Discharge</h2><h3 id="m0">testsLogBox.LoginTests.LoginPracticeTests#shouldLoginWithValidUsernameValidPassword</h3><table class="result"><tr><th>Messages</th></tr><tr><td><div class="messages">Logged in with a valid username and valid password into Expected URL = https://qa.logbox.co.za/premium/#/</div></td></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
