+++ LogBox20220513/test-output/junitreports/TEST-testsLogBox.HomePageTests.HomePageTests.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="YolandeH-Dell2" failures="0" tests="1" name="testsLogBox.HomePageTests.HomePageTests" time="9.882" errors="0" timestamp="01 Jun 2022 06:33:46 GMT" skipped="0">
+  <testcase classname="testsLogBox.HomePageTests.HomePageTests" name="shouldSearchForPatient" time="9.882"/>
+</testsuite> <!-- testsLogBox.HomePageTests.HomePageTests -->
