+++ LogBox20220513/test-output/junitreports/TEST-testsLogBox.LoginTests.LoginPracticeTests.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="YolandeH-Dell2" failures="0" tests="1" name="testsLogBox.LoginTests.LoginPracticeTests" time="3.195" errors="0" timestamp="01 Jun 2022 06:44:03 GMT" skipped="0">
+  <testcase classname="testsLogBox.LoginTests.LoginPracticeTests" name="shouldLoginWithValidUsernameValidPassword" time="3.195"/>
+</testsuite> <!-- testsLogBox.LoginTests.LoginPracticeTests -->
