+++ LogBox20220513/test-output/old/Charlotte/Pre-admission, Admission and Discharge.properties
@@ -1,0 +1,1 @@
+[SuiteResult context=Pre-admission, Admission and Discharge]
