+++ LogBox20220513/test-output/old/Charlotte/classes.html
@@ -4,14 +4,14 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>testsLogBox.HomePageTests.HomePageTests</td>
+<td>testsLogBox.LoginTests.LoginPracticeTests</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>shouldSearchForPatient</td>
+<td>shouldLoginWithValidUsernameValidPassword</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeClass</td>
