+++ LogBox20220513/test-output/old/Charlotte/main.html
@@ -1,0 +1,2 @@
+<html><head><title>Results for Charlotte</title></head>
+<body>Select a result on the left-hand pane.</body></html>
