+++ LogBox20220513/test-output/old/Charlotte/methods.html
@@ -1,0 +1,6 @@
+<h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Charlotte</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
+<table border="1">
+<tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
+<tr bgcolor="7abbb4">  <td>22/06/01 08:44:00</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="LoginPracticeTests.shouldLoginWithValidUsernameValidPassword()[pri:0, instance:testsLogBox.LoginTests.LoginPracticeTests@81ff872]">shouldLoginWithValidUsernameValidPassword</td> 
+  <td>main@1112075608</td>   <td></td> </tr>
+</table>
