+++ LogBox20220513/test-output/old/Charlotte/reporter-output.html
@@ -1,0 +1,2 @@
+<h2>Reporter output</h2><table><tr><td>Logged in with a valid username and valid password into Expected URL = https://qa.logbox.co.za/premium/#/</td></tr>
+</table>
