+++ LogBox20220513/test-output/old/Charlotte/testng.xml.html
@@ -1,0 +1,1 @@
+<html><head><title>testng.xml for Charlotte</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"http://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="Charlotte"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="Pre-admission,&nbsp;Admission&nbsp;and&nbsp;Discharge"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;packages&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;package&nbsp;name="testsLogBox.LoginTests"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/packages&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="testsLogBox.LoginTests.LoginPracticeTests"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;Pre-admission,&nbsp;Admission&nbsp;and&nbsp;Discharge&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;Charlotte&nbsp;--&gt;<br/></tt></body></html>
