+++ LogBox20220513/test-output/old/Default suite/methods.html
@@ -1,6 +1,6 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="a5a2f5">  <td>22/05/28 12:47:18</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="HomePageTests.shouldSearchForPatient()[pri:0, instance:testsLogBox.HomePageTests@7fcff1b9]">shouldSearchForPatient</td> 
-  <td>main@1300132917</td>   <td></td> </tr>
+<tr bgcolor="a5a2f5">  <td>22/06/01 08:33:36</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="HomePageTests.shouldSearchForPatient()[pri:0, instance:testsLogBox.HomePageTests.HomePageTests@40021799]">shouldSearchForPatient</td> 
+  <td>main@329481770</td>   <td></td> </tr>
 </table>
