+++ LogBox20220513/test-output/testng-results.xml
@@ -1,17 +1,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <testng-results ignored="0" total="1" passed="1" failed="0" skipped="0">
   <reporter-output>
+    <line>
+      <![CDATA[Logged in with a valid username and valid password into Expected URL = https://qa.logbox.co.za/premium/#/]]>
+    </line>
   </reporter-output>
-  <suite started-at="2022-05-28T10:47:18Z" name="Default suite" finished-at="2022-05-28T10:47:21Z" duration-ms="3730">
+  <suite started-at="2022-06-01T06:44:00Z" name="Charlotte" finished-at="2022-06-01T06:44:03Z" duration-ms="3217">
     <groups>
     </groups>
-    <test started-at="2022-05-28T10:47:18Z" name="Default test" finished-at="2022-05-28T10:47:21Z" duration-ms="3730">
-      <class name="testsLogBox.HomePageTests">
-        <test-method signature="shouldSearchForPatient()[pri:0, instance:testsLogBox.HomePageTests@7fcff1b9]" started-at="2022-05-28T10:47:18Z" name="shouldSearchForPatient" finished-at="2022-05-28T10:47:21Z" duration-ms="3710" status="PASS">
+    <test started-at="2022-06-01T06:44:00Z" name="Pre-admission, Admission and Discharge" finished-at="2022-06-01T06:44:03Z" duration-ms="3217">
+      <class name="testsLogBox.LoginTests.LoginPracticeTests">
+        <test-method signature="shouldLoginWithValidUsernameValidPassword()[pri:0, instance:testsLogBox.LoginTests.LoginPracticeTests@81ff872]" started-at="2022-06-01T06:44:00Z" name="shouldLoginWithValidUsernameValidPassword" finished-at="2022-06-01T06:44:03Z" duration-ms="3195" status="PASS">
           <reporter-output>
+            <line>
+              <![CDATA[Logged in with a valid username and valid password into Expected URL = https://qa.logbox.co.za/premium/#/]]>
+            </line>
           </reporter-output>
-        </test-method> <!-- shouldSearchForPatient -->
-      </class> <!-- testsLogBox.HomePageTests -->
-    </test> <!-- Default test -->
-  </suite> <!-- Default suite -->
+        </test-method> <!-- shouldLoginWithValidUsernameValidPassword -->
+      </class> <!-- testsLogBox.LoginTests.LoginPracticeTests -->
+    </test> <!-- Pre-admission, Admission and Discharge -->
+  </suite> <!-- Charlotte -->
 </testng-results>
