+++ LogBox20220513/testng.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="Charlotte">
+	<listeners>
+		<listener class-name="listeners.Listeners" />
+	</listeners>
+
+	<test thread-count="5" name="Pre-admission, Admission and Discharge">
+		<packages>
+			<package name="testsLogBox.LoginTests" />
+		</packages>
+	</test> <!-- Test -->
+</suite> <!-- Suite -->
